Support redis cache,the cache idea from nopcommerce,optimize the paging stored procedure(but can't execute mysql stored procedure in vs,can be execute by navicat),adjust content of views

+++ DapperRepository.Services/Mssql/Customers/CustomerService.cs
@@ -1,5 +1,8 @@
 ﻿using System;
 using System.Collections.Generic;
+using DapperRepository.Core;
+using DapperRepository.Core.Cache;
+using DapperRepository.Core.Constants;
 using DapperRepository.Core.Domain.Customers;
 using DapperRepository.Data.Repositories.BaseInterfaces;
 using DapperRepository.Services.BaseInterfaces;
@@ -9,10 +12,12 @@ namespace DapperRepository.Services.Mssql.Customers
     public class CustomerService : ICustomerService, IMssqlService
     {
         private readonly ICustomerRepository _repository;
+        private readonly ICacheManager _cacheManager;
 
-        public CustomerService(ICustomerRepository repository)
+        public CustomerService(ICustomerRepository repository, ICacheManager cacheManager)
         {
             _repository = repository;
+            _cacheManager = cacheManager;
         }
 
         public Customer GetCustomerById(int customerId)
@@ -31,9 +36,15 @@ namespace DapperRepository.Services.Mssql.Customers
             return _repository.GetCustomerBy(id);
         }
 
-        public int InsertList(out long time, List<Customer> customers)
+        public int InsertList(out long time, List<Customer> customers, int roleId)
         {
-            return _repository.InsertList(out time, customers);
+            var result = _repository.InsertList(out time, customers, roleId);
+            if (result > 0)
+            {
+                _cacheManager.RemoveByPattern(string.Format(CustomerDefaults.CustomerCountPatternCacheKey, ConnKeyConstants.Mssql));
+            }
+
+            return result;
         }
 
         public IEnumerable<CustomerDtoModel> GetAllCustomers()
@@ -41,9 +52,25 @@ namespace DapperRepository.Services.Mssql.Customers
             return _repository.GetAllCustomers();
         }
 
-        public IEnumerable<CustomerDtoModel> GetPagedCustomers(out int totalCount, int pageIndex = 0, int pageSize = int.MaxValue, bool useStoredProcedure = false)
+        public IEnumerable<CustomerDtoModelForPage> GetPagedCustomers(out int totalCount, string username = "", string email = "", int pageIndex = 0, int pageSize = int.MaxValue, bool useStoredProcedure = false)
         {
-            return _repository.GetPagedCustomers(out totalCount, pageIndex, pageSize, useStoredProcedure);
+            int total;
+
+            if (string.IsNullOrEmpty(username) && string.IsNullOrEmpty(email))
+            {
+                // 缓存无搜索条件的总记录数
+                total = _cacheManager.Get(string.Format(CustomerDefaults.CustomerTotalCountCacheKey, ConnKeyConstants.Mssql), 1440, () => _repository.GetCustomerCount());
+            }
+            else
+            {
+                string filterKey = string.Format(CustomerDefaults.CustomerFilterCountCacheKey, ConnKeyConstants.Mssql, CommonHelper.GetHashString(username + email));
+
+                total = _cacheManager.Get(filterKey, 1440, () => _repository.GetCustomerCount(username, email));
+            }
+
+            totalCount = total;
+
+            return _repository.GetPagedCustomers(total, username, email, pageIndex, pageSize, useStoredProcedure);
         }
 
         public int InsertCustomer(Customer customer, int roleId)
@@ -51,20 +78,35 @@ namespace DapperRepository.Services.Mssql.Customers
             if (customer == null)
                 throw new ArgumentNullException("customer");
 
-            return _repository.InsertCustomer(customer, roleId);
+            var result = _repository.InsertCustomer(customer, roleId);
+            if (result > 0)
+            {
+                _cacheManager.RemoveByPattern(string.Format(CustomerDefaults.CustomerCountPatternCacheKey, ConnKeyConstants.Mssql));
+            }
+
+            return result;
         }
 
         public int UpdateCustomer(Customer customer, int roleId)
         {
             if (customer == null)
                 throw new ArgumentNullException("customer");
-
             return _repository.UpdateCustomer(customer, roleId);
         }
 
         public bool DeleteCustomer(Customer customer)
         {
-            return _repository.Delete(customer.Id);
+            if (customer == null)
+                throw new ArgumentNullException("customer");
+
+            var result = _repository.Delete(customer.Id);
+
+            if (result)
+            {
+                _cacheManager.RemoveByPattern(string.Format(CustomerDefaults.CustomerCountPatternCacheKey, ConnKeyConstants.Mssql));
+            }
+
+            return result;
         }
     }
 }
